# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DermaAI - Skin Cancer Detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ConditionCount {
  condition: string;
  count: number;
  color: string;
}

export interface RiskFactors {
  /** @nullable */
  sunExposure?: string | null;
  /** @nullable */
  skinType?: string | null;
  /** @nullable */
  familyHistory?: boolean | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  immuneCondition?: boolean | null;
}

export interface RiskDataPoint {
  month: string;
  riskScore: number;
}

export interface Analysis {
  id: number;
  patientId: string;
  patientName: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  condition?: string;
  /** @nullable */
  conditionDetails?: string | null;
  prediction: string;
  riskLevel: string;
  confidenceScore: number;
  /** @nullable */
  cancerType?: string | null;
  /** @nullable */
  abcdeScore?: string | null;
  /** @nullable */
  lesionArea?: string | null;
  /** @nullable */
  recommendations?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  segmentedImageUrl?: string | null;
  /** @nullable */
  heatmapImageUrl?: string | null;
  riskFactors?: RiskFactors;
  riskProgressData?: RiskDataPoint[];
  explainableAiReasons?: string[];
  /** @nullable */
  reportId?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalAnalyses: number;
  highRiskCases: number;
  doctorsOnline: number;
  appointments: number;
  totalPatients: number;
  aiAccuracyRate: number;
  analysesToday: number;
  conditionDistribution: ConditionCount[];
  recentActivity: Analysis[];
}

export interface AnalysisInput {
  patientName: string;
  age?: number;
  gender?: string;
  conditionDetails?: string;
  imageData?: string;
  riskFactors?: RiskFactors;
}

export interface ImageUploadInput {
  patientName: string;
  conditionDetails?: string;
  imageData: string;
  age?: number;
  gender?: string;
  riskFactors?: RiskFactors;
}

export interface SimilarCase {
  id: number;
  patientName: string;
  prediction: string;
  riskLevel: string;
  confidenceScore: number;
  outcome?: string;
  /** @nullable */
  imageUrl?: string | null;
  date: string;
}

export interface AnalysisResult {
  analysisId: number;
  prediction: string;
  riskLevel: string;
  confidenceScore: number;
  cancerType?: string;
  abcdeScore?: string;
  lesionArea?: string;
  recommendations: string;
  imageUrl?: string;
  segmentedImageUrl?: string;
  heatmapImageUrl?: string;
  riskProgressData?: RiskDataPoint[];
  explainableAiReasons?: string[];
  similarCases?: SimilarCase[];
  reportId?: string;
  isHighRisk: boolean;
}

export interface ReportInfo {
  reportId: string;
  downloadUrl: string;
  qrCodeData: string;
  generatedAt: string;
}

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  hospital: string;
  city: string;
  country: string;
  experience: number;
  rating: number;
  phone?: string;
  email?: string;
  isOnline: boolean;
  /** @nullable */
  avatar?: string | null;
}

export interface ChatMessageInput {
  message: string;
  /** @nullable */
  analysisId?: number | null;
}

export interface ChatResponse {
  id: number;
  response: string;
  timestamp: string;
}

export interface ChatMessage {
  id: number;
  role: string;
  content: string;
  timestamp: string;
}

export type ListAnalysesParams = {
search?: string;
patientId?: string;
dateFrom?: string;
dateTo?: string;
};

export type ListDoctorsParams = {
city?: string;
};