# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DermaAI - Skin Cancer Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskDataPoint } from './riskDataPoint';
import type { RiskFactors } from './riskFactors';

export interface Analysis {
  id: number;
  patientId: string;
  patientName: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  condition?: string;
  /** @nullable */
  conditionDetails?: string | null;
  prediction: string;
  riskLevel: string;
  confidenceScore: number;
  /** @nullable */
  cancerType?: string | null;
  /** @nullable */
  abcdeScore?: string | null;
  /** @nullable */
  lesionArea?: string | null;
  /** @nullable */
  recommendations?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  segmentedImageUrl?: string | null;
  /** @nullable */
  heatmapImageUrl?: string | null;
  riskFactors?: RiskFactors;
  riskProgressData?: RiskDataPoint[];
  explainableAiReasons?: string[];
  /** @nullable */
  reportId?: string | null;
  createdAt: string;
}