# lib/api-zod/src/generated/types/analysisInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DermaAI - Skin Cancer Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskFactors } from './riskFactors';

export interface AnalysisInput {
  patientName: string;
  age?: number;
  gender?: string;
  conditionDetails?: string;
  imageData?: string;
  riskFactors?: RiskFactors;
}