# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DermaAI - Skin Cancer Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskDataPoint } from './riskDataPoint';
import type { SimilarCase } from './similarCase';

export interface AnalysisResult {
  analysisId: number;
  prediction: string;
  riskLevel: string;
  confidenceScore: number;
  cancerType?: string;
  abcdeScore?: string;
  lesionArea?: string;
  recommendations: string;
  imageUrl?: string;
  segmentedImageUrl?: string;
  heatmapImageUrl?: string;
  riskProgressData?: RiskDataPoint[];
  explainableAiReasons?: string[];
  similarCases?: SimilarCase[];
  reportId?: string;
  isHighRisk: boolean;
}