# lib/api-zod/src/generated/types/imageUploadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DermaAI - Skin Cancer Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskFactors } from './riskFactors';

export interface ImageUploadInput {
  patientName: string;
  conditionDetails?: string;
  imageData: string;
  age?: number;
  gender?: string;
  riskFactors?: RiskFactors;
}