# lib/api-zod/src/generated/types/similarCase.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DermaAI - Skin Cancer Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface SimilarCase {
  id: number;
  patientName: string;
  prediction: string;
  riskLevel: string;
  confidenceScore: number;
  outcome?: string;
  /** @nullable */
  imageUrl?: string | null;
  date: string;
}